requirements.txt
requests==2.31.0
websocket-client==1.6.4
pandas==2.1.4
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.8
anthropic==0.8.1
sqlalchemy==2.0.23
schedule==1.2.0
google-generativeai==0.3.2